Added missing getters for window caption and update_interval

+++ Gosu/Window.h
@@ -22,7 +22,7 @@ void Gosu_Window_needs_cursor(Gosu_Window* window, void function());
 int Gosu_Window_width(Gosu_Window* window);
 int Gosu_Window_height(Gosu_Window* window);
 int Gosu_Window_fullscreen(Gosu_Window* window);
-int Gosu_Window_update_interval(Gosu_Window* window);
+double Gosu_Window_update_interval(Gosu_Window* window);
 int Gosu_Window_resizable(Gosu_Window* window);
 const char* Gosu_Window_caption(Gosu_Window* window);
